LayoutTests/http/tests/from-origin/fetch-from-origin-same-site-blocked-expected.txt

CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/from-origin/resources/fetch.php?fromOrigin=same-site. Cancelled load because it violates the resource's From-Origin response header.
Tests that a cross-origin fetch load fails if the server blocks it with a 'From-Origin: same-site' response header, even though the server includes an Access-Control-Allow-Origin response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Fetch blocked. TypeError: Cancelled load because it violates the resource's From-Origin response header.
PASS successfullyParsed is true

TEST COMPLETE
